+++ ChangeLog.md
@@ -17,6 +17,7 @@
     - Obsoletize Modal.setFooter() replaced with Modal.set() (todo #35)
     - Obsoletize Modal.setClasses() replaced with Modal.set() (todo #32)
     - Obsoletize Modal.setBody() replaced with Modal.set() (todo #34)
+    - Define classesContent, classesBody, classesFooter and classesHeader to add classes to respective dialog parts (todo #29)
 
 ### 1.6.0